Add "by" argument to supply grouping variable.
Replace the "group1" and "group2" arguments in compare_means by a single
"by" argument.  The group arguments can lead to subtle errors in cases
like the following:

    compare_means(vars, d$study == 1, d$study == 2, d[d$area == 1, ]))

The "group1" and "group2" arguments refer to the full-sized dataset
before subsetting.  They are incompatible in terms of length with the
partial dataset supplied as the "data" argument.  Since the function
call looks fine at first glance this is an easy mistake to make.

The new "by" argument avoids this problem because it names a variable in
the "data" argument.  The price to pay is that there must be a two-level
categorical variable in the dataset that defines the groups for the
comparison of means.  If such a variable does not exists, it has to be
created prior to calling compare_means.

Provided that the dataset d has a "study" variable with levels 1 and 2,
and possibly some missing values, the above function call using the new
"by" argument can be written as:

    compare_means(vars, by = "study", d[d$area == 1, ])